+++ app/src/main/res/layout/activity_main.xml
@@ -4,19 +4,21 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="#FFFFFF"
     tools:context=".MainActivity">
 
     <LinearLayout
-        android:orientation="vertical"
+        android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_width="match_parent">
+        android:orientation="vertical"
+        tools:layout_editor_absoluteX="0dp"
+        tools:layout_editor_absoluteY="0dp">
 
 
         <RelativeLayout
-            android:layout_height="60dp"
             android:layout_width="match_parent"
-            android:background="#ffffffff"
-            >
+            android:layout_height="60dp"
+            android:background="#ffffffff">
 
             <Button
                 android:id="@+id/monthPrevious"
@@ -34,6 +36,7 @@
                 android:layout_centerInParent="true"
                 android:gravity="center_horizontal"
                 android:text="MonthView"
+                android:textColor="@color/black"
                 android:textSize="30dp"
                 android:textStyle="bold" />
 
@@ -41,11 +44,10 @@
                 android:id="@+id/monthNext"
                 android:layout_width="20dp"
                 android:layout_height="45dp"
+                android:layout_alignParentRight="true"
                 android:layout_margin="13dp"
                 android:background="@drawable/forward"
-                android:gravity="center_horizontal"
-                android:layout_alignParentRight="true"
-                />
+                android:gravity="center_horizontal" />
 
         </RelativeLayout>
 
@@ -65,7 +67,7 @@
         <GridView
             android:id="@+id/monthView"
             android:layout_width="match_parent"
-            android:layout_height="520dp"
+            android:layout_height="440dp"
             android:background="#ffbbbbbb"
             android:horizontalSpacing="1dp"
             android:numColumns="7"
@@ -80,16 +82,19 @@
             tools:visibility="visible">
 
             <Button
-                android:id="@+id/mypeel"
+                android:id="@+id/myPeel"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_centerInParent="true"
+                android:background="#FFFFFF"
                 android:gravity="center_horizontal"
                 android:text="내 약"
+                android:textColor="@color/ic_launcher_background"
+
                 android:textSize="30dp"
                 android:textStyle="bold"
-
-                 />
+                app:iconTint="#FFFFFF"
+                app:strokeColor="#FFFFFF" />
 
 
         </RelativeLayout>
